backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
bcrypt==4.1.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
gunicorn==21.2.0 
pip install pyodbc sqlalchemy